enumutil: add Enum class for wrapping rust enums
Enum values received from libvirt are usually mapped to Rust enum
constants. The libvirt driver that the app is talking to, however,
might be returning values for which there is no Rust mapping yet.
This is typical where new libvirt introduces new enum constants,
but the app is using an older libvirt version, or simply because
the latest Rust binding is not yet feature complete.

When exposing enums in the API, it is important that we do not
throw away any information that we received from libvirt, so we
need a way to preserve unknown enum values.

To achieve this, a new generic type is introduced that wraps the
root Rust enum type:

  pub enum Enum<E, I> {
    Known(E),
    Unknown(I),
  }

Consider a Rust enum

  enum ErrorLevel {
    None,
    Warning,
    Error,
  }

which maps to the virErrorLevel C type, which may either be a
plain integer type with CPP constants, or a real C enum.
Either way, we can now define a Rust enum wrapper:

  pub type ErrorLevelEnum = Enum<ErrorLevel, sys::virErrorLevel>

The Rust APIs / structs that are to be populated with enum values
received from libvirt should all use the wrapper, ErrorLevelEnum.
When mapping the libvirt value there are two options, either using
a well known Rust enum identifier:

  let e: ErrorLevelEnum = Enum::Known<Warning>;

Or using an unknown C value

  let e: ErrorLevelEnum = Enum::Unknown(42);

Or if we don't know if the C value is known or unknown:

  let e: ErrorLevelEnum = from_raw(2);

A caller that needs to be able to interpret the enum values has
to decide their policy for unknown values.

The lazy option is to panic by calling unwrap(), which is suitable
if you are sure no unexpected values are present, or in throwaway
tools not needing full error handling:

  let l: ErrorLevel = err.level().unwrap()

A more intelligent option is to squash unknown values to a known
value that will give good semantics. eg treat unknown error levels
as being error

  let l: ErrorLevel = err.level().unwrap_or(ErrorLevel::Error)

Finally the most explicit option is to manually match

  let l: ErrorLevel = match err.level() {
    Known(l) => ...
    Unknown(val) => ...
  };

There are various other helper methods on the Enum type, inspired
by similar methods on the Option and Result rust types.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>