enumutil: drop support for default mappings from raw value
The 'Last' field in enums throws away the raw value so it is not
especially useful as a distinct value. Furthermore, Rust has a
standard way of indicating missing data via the 'Option' type.

Change 'from_raw' to return 'Option<TheEnumType>', instead of
'TheEnumType'.

In the error class we then unwrap the enum option, rewriting
the missing data to an different known enum field.

This is a slight behaviour change for the ErrorDomain and ErrorNumber
types as applications can no longer distinguish a missing value. This
will be addressed in a following commit.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>