src: make many more methods "void"
Most libvirt methods which return an integer return -1 on error
and 0 on success. We turn -1 into an Error within the Result
object, but in many cases we also return the value 0 as a u32.
The latter is pointless as it conveys no information. The return
value should essentially be "void", which we represent by ().

A handful of methods are left returning u32, since they return
counts or sizes.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>